Fix: Add user fields fetched at runtime to SSE dataset CS filter * Move CS User field handling logic to email.service
(cherry picked from commit aefc04a1aef438fa36c6492faee5e15725205603)
